Update readCSV.py (#1)
* Add files via upload

* Update readCSV.py

* Update readCSV.py

Added all the required fields. Now working on to the next step.

* Update readCSV.py

Just updated the program as per specifications of baba when I just started on this simulation program. The algorithm part to buy and sell automatically is yet to be implemented. So for now just the first part and initial frameworking